- icons.md was missing, found a copy from previous commit and updated the instructions, which now include steps on renaming the icons based on taxonomies - added a default zipfile from fontello so users can start modifying from that one - updated code to get css from icon.css to fontello.css, so the user just has to replace the zipped folders instead of finding the fontello.css file and renaming it to icon.css.
