+++ components/Home/Home.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import s from './Home.css'
 import CategoryList from '../CategoryList'
-import icons from '../../icons/css/icons.css'
+import icons from '../../icons/css/fontello.css'
 
 const Home = (props) => (
   <div>
