+++ icons/css/fontello.css
@@ -1,0 +1,80 @@
+@font-face {
+  font-family: 'fontello';
+  src: url('../font/fontello.eot?88085213');
+  src: url('../font/fontello.eot?88085213#iefix') format('embedded-opentype'),
+       url('../font/fontello.woff2?88085213') format('woff2'),
+       url('../font/fontello.woff?88085213') format('woff'),
+       url('../font/fontello.ttf?88085213') format('truetype'),
+       url('../font/fontello.svg?88085213#fontello') format('svg');
+  font-weight: normal;
+  font-style: normal;
+}
+/* Chrome hack: SVG is rendered more smooth in Windozze. 100% magic, uncomment if you need it. */
+/* Note, that will break hinting! In other OS-es font will be not as sharp as it could be */
+/*
+@media screen and (-webkit-min-device-pixel-ratio:0) {
+  @font-face {
+    font-family: 'fontello';
+    src: url('../font/fontello.svg?88085213#fontello') format('svg');
+  }
+}
+*/
+ 
+ [class^="icon-"]:before, [class*=" icon-"]:before {
+  font-family: "fontello";
+  font-style: normal;
+  font-weight: normal;
+  speak: none;
+ 
+  display: inline-block;
+  text-decoration: inherit;
+  width: 1em;
+  margin-right: .2em;
+  text-align: center;
+  /* opacity: .8; */
+ 
+  /* For safety - reset parent styles, that can break glyph codes*/
+  font-variant: normal;
+  text-transform: none;
+ 
+  /* fix buttons height, for twitter bootstrap */
+  line-height: 1em;
+ 
+  /* Animation center compensation - margins should be symmetric */
+  /* remove if not needed */
+  margin-left: .2em;
+ 
+  /* you can be more comfortable with increased icons size */
+  /* font-size: 120%; */
+ 
+  /* Font smoothing. That was taken from TWBS */
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+ 
+  /* Uncomment for 3D effect */
+  /* text-shadow: 1px 1px 1px rgba(127, 127, 127, 0.3); */
+}
+ 
+.icon-housing:before { content: '\e800'; } /* '' */
+.icon-hygiene:before { content: '\e801'; } /* '' */
+.icon-down-open:before { content: '\e802'; } /* '' */
+.icon-left-open:before { content: '\e803'; } /* '' */
+.icon-right-open:before { content: '\e804'; } /* '' */
+.icon-angle-double-right:before { content: '\e805'; } /* '' */
+.icon-technology:before { content: '\e806'; } /* '' */
+.icon-check:before { content: '\e807'; } /* '' */
+.icon-food:before { content: '\e808'; } /* '' */
+.icon-compass:before { content: '\e80a'; } /* '' */
+.icon-phone:before { content: '\e80b'; } /* '' */
+.icon-right-open-1:before { content: '\e80c'; } /* '' */
+.icon-droplet:before { content: '\e80d'; } /* '' */
+.icon-right-open-outline:before { content: '\e80e'; } /* '' */
+.icon-plus:before { content: '\e80f'; } /* '' */
+.icon-right-open-2:before { content: '\e810'; } /* '' */
+.icon-link:before { content: '\e811'; } /* '' */
+.icon-user:before { content: '\e812'; } /* '' */
+.icon-user-woman:before { content: '\e813'; } /* '' */
+.icon-user-pair:before { content: '\e814'; } /* '' */
+.icon-ok:before { content: '\e817'; } /* '' */
+.icon-check-empty:before { content: '\f096'; } /* '' */
+.icon-medical:before { content: '\f0fa'; } /* '' */
