+++ lib/taxonomies.js
@@ -1,4 +1,4 @@
-import icons from '../icons/css/icons.css'
+import icons from '../icons/css/fontello.css'
 import { capitalize } from './stringHelpers'
 
 export const getIcon = (taxonomy) => (
